moved ServiceNode class out into a separate file
